Refactor: Project structure
