+++ src/helpers/assetLoader.js
